try to show only related user-jobs

+++ app/Http/Controllers/JobController.php
@@ -6,6 +6,7 @@ use App\Models\Job;
 use App\Http\Requests\StoreJobRequest;
 use App\Http\Requests\UpdateJobRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class JobController extends Controller
 {
@@ -16,14 +17,24 @@ class JobController extends Controller
      */
     public function index()
     {
-        $jobs = Job::orderBy('time', 'desc')->get();
-        return response()->json($jobs);
+        if (Auth::check()) {
+            $user = Auth::user();
+            $jobs = Job::where('user_id', $user->id)->orderBy('time', 'desc')->get();
+            return response()->json($jobs);
+        } else {
+
+            return response()->json(['error' => 'Unauthenticated'], 401);
+        }
     }
 
+    //continue: show, create and update only with connection to the user
     public function jobsByCategory(Request $request)
     {
         $categoryId = $request->route('category_id');
-        $jobs = Job::where('category_id', $categoryId)->orderBy('time', 'desc')->get();
+        $jobs = Job::where('user_id', Auth::id())
+            ->where('category_id', $categoryId)
+            ->orderBy('time', 'desc')
+            ->get();
 
         return response()->json($jobs);
     }
@@ -48,6 +59,8 @@ class JobController extends Controller
     {
         $valData = $request->validated();
         $valData['category_id'] = $request->input('category_id');
+
+        $valData['user_id'] = Auth::id();
         $job = Job::create($valData);
         return response()->json(['job' => $job, 'message' => 'Job Added Successfully']);
     }
@@ -60,6 +73,9 @@ class JobController extends Controller
      */
     public function show(Job $job)
     {
+        if ($job->user_id != Auth::id()) {
+            return response()->json(['error' => 'You do not have permission to view this job.'], 403);
+        }
         $job = Job::find($job->id);
         return response()->json($job);
     }
@@ -87,6 +103,10 @@ class JobController extends Controller
         $job = Job::find($job->id);
         $valData = $request->validated();
         $valData['category_id'] = $request->input('category_id');
+
+        if ($job->user_id != Auth::id()) {
+            return response()->json(['error' => 'You do not have permission to update this job.'], 403);
+        }
         $job->update($valData);
         return response()->json(['job' => $job, 'message' => 'Job Updated Successfully']);
     }
@@ -100,6 +120,10 @@ class JobController extends Controller
     public function delete(Job $job)
     {
         $job = Job::find($job->id);
+
+        if ($job->user_id != Auth::id()) {
+            return response()->json(['error' => 'You do not have permission to delete this job.'], 403);
+        }
         $job->delete();
         return response()->json(['message' => 'Job deleted Successfully']);
     }
